viewer/requirements.txt
opencv-python>=4.9.0
Pillow>=10.0.0
screeninfo>=0.8.1
watchdog>=4.0.0
pymediainfo>=6.0.1
numpy>=1.26.0